Move RPC method to isolated module class
