+++ evm/rpc/modules/__init__.py
@@ -1,0 +1,3 @@
+from .main import RPCModule  # noqa: F401
+
+from .eth import Eth  # noqa: F401
