+++ evm/rpc/modules/main.py
@@ -1,0 +1,5 @@
+
+
+class RPCModule:
+    def __init__(self, chain):
+        self._chain = chain
